move DsEngine gst_config.py was getting too long
