+++ applications/smart-distancing/smart_distancing/detectors/deepstream/__init__.py
@@ -13,11 +13,13 @@ from gi.repository import (
 from smart_distancing.detectors.deepstream._pyds import *
 from smart_distancing.detectors.deepstream._ds_config import *
 from smart_distancing.detectors.deepstream._ds_engine import *
+from smart_distancing.detectors.deepstream._ds_engine import *
 from smart_distancing.detectors.deepstream._detectors import *
 
 __all__ = [
     'DsConfig',  # _ds_config.py
     'DsDetector',  # _detectors.py
+    'DsEngine',  # _ds_engine.py
     'ElemConfig',  # _ds_config.py
     'frame_meta_iterator',  # _ds_engine.py
     'GstConfig',  # _ds_config.py
